chart rating to double

+++ src/main/java/com/example/shopBackend/review/Chart.java
@@ -7,6 +7,6 @@ package com.example.shopBackend.review;
  */
 @SuppressWarnings("unused")
 public interface Chart {
-	int getRating();
+	double getRating();
 	int getCount();
 }
